Added service interfaces to repository services
